Replaced int types with doubles
Changed lines of int position to double position

+++ src/main/cpp/subsystems/Elevator.cpp
@@ -57,7 +57,7 @@ void Elevator::SetElevator(double speed) {
           mRightElevator.Set(ControlMode::PercentOutput,speed);
 }
 
-void Elevator::SetElevatorPosition(int position){
+void Elevator::SetElevatorPosition(double position){
    mRightElevator.Set(ControlMode::MotionMagic, position);
 }
